+++ Sources/BraintreeCore/Analytics/BTAnalyticsService.swift
@@ -41,21 +41,21 @@ class BTAnalyticsService: Equatable {
     /// Sends analytics event to https://api.paypal.com/v1/tracking/batch/events/ via a background task.
     /// - Parameters:
     ///   - eventName: Name of analytic event.
-    ///   - isVaultRequest: Optional. If the Venmo or PayPal request is being vaulted.
     ///   - correlationID: Optional. CorrelationID associated with the checkout session.
     ///   - endpoint: Optional. The endpoint of the API request send during networking requests.
     ///   - endTime: Optional. The end time of the roundtrip networking request.
     ///   - errorDescription: Optional. Full error description returned to merchant.
+    ///   - isVaultRequest: Optional. If the Venmo or PayPal request is being vaulted.
     ///   - linkType: Optional. The type of link the SDK will be handling, currently deeplink or universal.
     ///   - payPalContextID: Optional. PayPal Context ID associated with the checkout session.
     ///   - startTime: Optional. The start time of the networking request.
     func sendAnalyticsEvent(
         _ eventName: String,
-        isVaultRequest: Bool? = nil,
         correlationID: String? = nil,
         endpoint: String? = nil,
         endTime: Int? = nil,
         errorDescription: String? = nil,
+        isVaultRequest: Bool? = nil,
         linkType: String? = nil,
         payPalContextID: String? = nil,
         startTime: Int? = nil
@@ -67,6 +67,7 @@ class BTAnalyticsService: Equatable {
                 endpoint: endpoint,
                 endTime: endTime,
                 errorDescription: errorDescription,
+                isVaultRequest: isVaultRequest,
                 linkType: linkType,
                 payPalContextID: payPalContextID,
                 startTime: startTime
@@ -77,11 +78,11 @@ class BTAnalyticsService: Equatable {
     /// Exposed to be able to execute this function synchronously in unit tests
     func performEventRequest(
         _ eventName: String,
-        isVaultRequest: Bool? = nil,
         correlationID: String? = nil,
         endpoint: String? = nil,
         endTime: Int? = nil,
         errorDescription: String? = nil,
+        isVaultRequest: Bool? = nil,
         linkType: String? = nil,
         payPalContextID: String? = nil,
         startTime: Int? = nil